Avoid a thundering herd ...
From the ticket:

    If we need to measure 90% of the relays in the network, we schedule
    ourselves to do so ASAP. Thus they will all need measuring again at
    roughly the same time. So every few days we'll do a ton of work, and
    otherwise do very little.

    We should spread out the load we create.

    Perhaps a better idea would be to have each relay's measurement expire
    at a random time X..Y days from now. After the first huge all-at-once
    scan, the ~entire network will next be measured slowly over the course
    of a few days.

This implements that idea. When we record the result from a relay's
measurement, we choose an expire time that is between 3 and 6 days in
the future. After a while, we will convert from a thundering herd of
relays needing a new measurement to slowly measuring the entire network
over 6 days.

I simulated this to make sure my intuition is correct. The script is at
the bottom of this commit message if you'd like to play with it. As DAYS
gets larger (i.e. the longer the sim runs), the printed range
approximates the max time to add (6).

State file version updated from 1 to 2 to add expire times.

closes #5

    #!/usr/bin/env python3
    import random
    N_OBJS = 700  # reasonable estimate of number of exit rlays
    DAYS = 100  # how long to run the sim
    ADD_MINMAX = (3, 6)  # how many days to add

    def main():
        objs = [0] * N_OBJS
        while objs[0] < DAYS:
            # get the obj with the smallest time at the front
            # because we are now at that time and are measuring it
            objs.sort()
            # add a random amount of time to when we should measure it
            # next
            objs[0] += random.uniform(*ADD_MINMAX)
        # sim is done. print stats
        objs.sort()
        minmax = (min(objs), max(objs))
        print(f'min/max: {minmax[0]}/{minmax[1]}')
        print(f'range: {minmax[1] - minmax[0]}')
        print(objs)

    if __name__ == '__main__':
        main()