Also scan relays who can reach our web IRC client ...
"Scan" for them means to a DNS "trick" to discover any additional IPs
the relay might use. Take the IPs from its descriptor, do a reverse DNS
lookup to get hostnames, and for each hostname, do a DNS lookup.

We actually do this DNS thing for all relays, because why not?

closes #7